requirements.txt
Django>=5.0.0,<5.1.0
Pillow>=10.0.0,<11.0.0
python-dateutil>=2.8.0,<3.0.0
python-decouple>=3.8
gunicorn>=21.2.0
openpyxl>=3.1.0,<4.0.0
reportlab>=4.0.0,<5.0.0
requests>=2.28.0
django-axes>=6.0.0
cryptography>=41.0.0
